Version 0.1.0
Version 0.1.0

Full working alpha with 4 Buttons.

This is seemingly in a fully working state as it sits. A full build
might be neccessary to go further with the code.

I need to try and figure out a way to make the functions more streamlined
and not as chaotic looking. Maybe grouping could do this, but I'm not sure
that wont just add complexity for the sake of simply making the code
easier to look at.

!!!TODO!!!
figure out flash storage and save elBtnState to that.
also need to make a hlBtnState and do the same for headlight function.